Rename xhci schedule header
