+++ rom/usb/pcixhci/xhci_hcd.c
@@ -18,7 +18,7 @@
 
 #include "uhwcmd.h"
 #include "xhciproto.h"
-#include "xhcichip_schedule.h"
+#include "xhci_schedule.h"
 
 #if defined(DEBUG) && defined(XHCI_LONGDEBUGNAK)
 #define XHCI_NAKTOSHIFT         (8)
@@ -4071,4 +4071,3 @@ void xhciFreeDeviceCtx(struct PCIController *hc,
 
     FreeMem(devCtx, sizeof(struct pciusbXHCIDevice));
 }
-
