+++ rom/usb/pcixhci/xhci_schedule.c
@@ -12,7 +12,7 @@
 #include <exec/memory.h>
 
 #include "uhwcmd.h"
-#include "xhcichip_schedule.h"
+#include "xhci_schedule.h"
 
 #ifdef base
 #undef base
@@ -356,4 +356,3 @@ WORD xhciQueuePayloadTRBs(struct PCIController *hc, struct IOUsbHWReq *ioreq,
 
     return queued;
 }
-
